[feat] 세부 모델 페이지 컴포넌트 분리
