+++ frontend/Idle/src/components/detailModelPage/detailModelSub/DetailModelBoxContainer.jsx
@@ -1,8 +1,9 @@
+import { styled } from "styled-components";
 import DetailModelBox from "./DetailModelBox"
 
 function DetailModelBoxContainer({ detailData, currentTab, setOptionsToBeRemoved, setWarningModalVisible }) {
     return (
-        <>
+        <StContainer>
             {detailData.map((item, idx) => (
                 <DetailModelBox
                     key={idx}
@@ -12,9 +13,15 @@ function DetailModelBoxContainer({ detailData, currentTab, setOptionsToBeRemoved
                     setModalVisible={setWarningModalVisible}
                 />
             ))}
-        </>
+        </StContainer>
 
     )
 }
 
 export default DetailModelBoxContainer
+
+const StContainer = styled.div`
+  display: inline-flex;
+  align-items: flex-start;
+  gap: 8px;
+`;
